backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
Pillow==10.2.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
drf-spectacular==0.27.1
django-import-export==3.3.5
openpyxl==3.1.2
pandas==2.2.0
boto3==1.34.25
django-guardian==2.4.0
channels==4.0.0
channels-redis==4.2.0
django-cleanup==8.1.0
django-simple-history==3.4.0
pytest-django==4.7.0
factory-boy==3.3.0
